Properly attach events on 2nd and subsequent renders of a view
This fixes the issue where events would work the first time a view was
shown, but not afterward.

Keeping the view *instances* in the app object requires this logic to handle
removing from the DOM as well as unhooking the events, and then hooking them
back up on subsequent renders.